Update README, refactoring and adds test infra
