+++ tests/minimal_init.lua
@@ -1,0 +1,19 @@
+local plenary_dir = os.getenv('PLENARY_DIR') or '/tmp/plenary.nvim'
+
+local is_directory = vim.fn.isdirectory(plenary_dir) ~= 0
+
+if not is_directory then
+  vim.fn.system({
+    'git',
+    'clone',
+    'https://github.com/nvim-lua/plenary.nvim',
+    plenary_dir,
+  })
+end
+
+vim.opt.rtp:append('.')
+vim.opt.rtp:append(plenary_dir)
+
+vim.cmd('runtime plugin/plenary.vim')
+
+require('plenary.busted')
